+++ src/main.jsx
@@ -1,10 +1,25 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './pages/HomePage'
+import Home from './pages/Home'
 import './index.css'
+
+const pages = [
+  {
+    id: "home",
+    name: "Home",
+  },
+  {
+    id: "about",
+    name: "About",
+  },
+  {
+    id: "portfolio",
+    name: "Portfolio",
+  },
+];
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <App />
+    <Home />
   </React.StrictMode>,
 )
