+++ src/pages/Home/style.scss
@@ -1,6 +1,6 @@
-article {
+#home {
     width: 100%;
-    height: 100%;
+    height: 100vh;
     display: flex;
     justify-content: center;
     align-items: center;
@@ -26,6 +26,7 @@ article {
                 font-weight: 100;
 
                 b {
+                    color: #2BCAB6;
                     font-weight: 400;
                 }
             }
@@ -42,6 +43,7 @@ article {
                 gap: 15px;
 
                 div {
+                    color: #2BCAB6;
                     font-weight: 400;
                 }
             }
